Use soft values in the ResourceFolderRepository cache
ResourceFolderRegistry contains a cache mapping from res folders to
their ResourceFolderRepository instances. This cache is eagerly
populated by a dumb mode task (PopulateCachesTask). However, since the
cached values were wrapped in weak references, they may be easily
reclaimed if they were not yet referenced from a composite repository
such as the ModuleResourceRepository. This leads to the repository being
re-created again, quite often from the EDT.

This CL makes the cache to wrap the values using SoftReferences. The
number of entries corresponds to the number of res folders in the
project, which is unlikely to be more than a 1000.

Bug: 123038839
Test: manually tested that the cache doesn't evict values eagerly.
Change-Id: I8bc87369e9414a7e297fcf48dd57b69928f1ace9