Add some improvements to async logic on palette.
As Sergey pointed out it is better to deal with immutable lists.
This CL makes that change.

Note that the method checkForRelevantDependencyChanges is
synchronized here such that the updates are always completed
fully. If multiple requests comes in (which they shouldn't)
then only one will execute at a time.

Also: added atomic reference and boolean to remove all the
warnings in this file.

Bug: 122373349
Test: Nothing additional to add.
It is better to deal with immutable lists when dealing with
multiple threads

Change-Id: Iee77ae410da0e610860e45105e3e9f6e2acca580