Fix hangs in Android Studio when stopping screen recording
In Android Studio we have the ability to record the device screen. When
start recording, we were sending shell commands to the device that can take
a while to run (e.g. "settings get system show_touches"). As we were doing
all of that from the UI thread, Android Studio would freeze if that command
took long enough. In some reports (e.g. crash/9b8775e7a6ec3063), the IDE
froze for almost 50 seconds.

To fix this issue, we need to make sure that all the shell commands are run
in a separate thread, backed by a ProgressIndicator that is displayed while
such operations are running in background.

Bug: 122376044
Test: ScreenRecorderActionTest still passing.
Change-Id: I55fe9cbe174b409f5bbe9ba1f5bec70c1f297baf