Avoid NetworkCache making unintended network calls
Some uses of GoogleMavenRepository run on the UI thread and could cause
blocks since GoogleMavenRepository might make network calls.
This CL is part of a workaround to avoid the calls on the UI thread
while a more comprehensive fix is done. The fix should split the
processing of the template rendering into smaller components so not
everything needs to happen under a read lock and in the UI thread.

Bug: 122113652
Test: Existing tests pass
Change-Id: Ibced6109628ee469da84df66dc31610da5d6c7f6