Remove SceneComponent.setModelUpdateAuthorized() API
By default myIsModelUpdateAuthorized is True, and only
LinearDragTarget set it as false.
When this attribute is true, setPositionTarget(),
setPosition(), setSize(), and setSizeTarget() always work
and ignore isFromModel argument.

When LinearDragTarget calles SceneComponent.setPosition(),
it set isFromModel attribute as false and setPosition()
still works. So actually myIsModelUpdateAuthorized does
nothing and can be removed.

After removing SceneComponent.setModelUpdateAuthorized(),
argument isFromModel is not needed for setPosition(),
setPositionTarget(), setSize(), and setSizeTargeT()
anymore.

This CL is for internal cleanup.

Bug: N/A
Test: Covered by existing tests.
Change-Id: Ic2ad1e14447b0c53db04055109aea83d67817202