Do not consume key events for completions and dropdown.
If no completions are shown, do not consume navigation keys for
completions.
Also if the drop down is not visible in a combo box, do not consume
the navigation key events.
Also do not consume PgUp and PgDn events in the table.

After this change PgUp and PgDn will scroll the properties panel,
without moving focus.

In this CL: Also rename all the KeyStrokes constants to use
uppercase.

Fixes: 122531954
Test: N/A
Change-Id: I8a7aaf08313d61a7dffdec4a298784db285b3880