Try to open existing project before re-importing
This updates the "open" action to look for a .idea folder. If present,
it will attempt to reopen the existing project before attempting to
perform a gradle import (which will overwrite any per-project settings).

In order to keep the "new window" prompt consistent and avoid repetition,
that logic has moved to a new method called promptToCloseIfNecessary which
is now invoked in both the open existing case and the import new case.

Change-Id: I0f3cb5ba96051fde27dedd0473c737ce3a35d85b
Test: Manually tested by opening an existing project with custom settings
Fixes: 78097962