Make AndroidRunConfigurationType dumb aware.
Currently, between when a build finishes and when deployment happens, there is
a large delay of 300ms-1s (dependent on machine specs). This is due to the Run
task being scheduled as a smart task. This change change Run to become a dumb
task, thus getting executed immediately and not wait for IDEA to sync.

Bug: 123369414
Test: N/A
Change-Id: I9bf66f43e821cc1d02dffb381ea67bd1aa2ddf55