Change calculation of anonymized project ID.
It uses the base path of the project
 1. To match the build system.
 2. To keep the ID stable across syncs and active variant changes. (b/123461345)

In the previous calculation, the anonymized id is generated based on
data from AndroidFacets. AndroidFacets are first time added to the
project on completion of the first successful sync.

Therefore even a simple session representing  "open project, close
project" may include two project ids, if the project is freshly created
or checked out from the VCS.

The default project is reported as DEFAULT.

The disadvantage of the new formulation is that a project checked out in
multiple directories by the same user will have different anonymized
IDs, and moving the project in to a different directory will change its
ID.

Change-Id: I3cf2e98f0f4547d145df7979e18563a17e9460ee
Fixes: 123461345
Test: Manually verified to be consistent. TODO(b/123519955): Automated tests.