NlModel#addComponents shouldn't make network operations
When adding components to the NlModel, we had a flow that would analyze
gradle dependencies by establishing network connections through
IdeGoogleMavenRepository. As such operations are blocking, they shouldn't
run on the UI thread, so we should wrap them in a ProgressIndicator using
a Task.Backgroundable that will run in the pooled executor.

Bug: 121239170
Test: Designer tests should pass
Change-Id: I52f95e3dba60c64a130a02acc071fe409f7e1c90