Don't prompt twice to open project in new window
This simplifies the logic that suppresses redundant gradle imports.
Rather than implementing an alternative "open" prompt when the .idea
is present, it just supresses the branch that peformms a gradle import.
This causes control to fall through to the general case at the end
of the method, which does the right thing.

Test: Manual: Enable the "Confirm window to open project in" setting.
      Open a project. Try all three options and confirm correct behavior
      and no double prompt. Repeat the test with multiple projects open,
      when the project being opened is one of the projects in a different
      window.
Fixes: 123552060

Change-Id: I95c0cd5b57c8a85f9bdc5212b5fda315b0c11096