Run DebugOnEmulatorTest within Bazel sandbox
DebugOnEmulatorTest is apparently able to run within the Bazel sandbox
now. This was not always the case, however. ADB used to be unable to run
within the Bazel sandbox because ADB tried to create the .android
directory under the user's home directory. When running in Bazel, the
user running the test is the nobody user, and ADB determines that the
path "/nonexistent" is the user's home directory. ADB would then try to
create the /nonexistent/.android directory, which caused ADB to fail to
start.  Without ADB, the test was unable to connect to an ADB server, so
the test would also fail.

This commit allows the DebugOnEmulatorTest to run within the sandbox
now. Additionally, this commit also addresses the testing error where
ADB tried to create the .android directory under a nonexistent
directory; it is possible the change to ADB or code managing ADB is
undone, which would cause ADB to write back to .android under the user's
home directory. Since this test is not a regression test for ADB nor
individual IDE components, this test should not be fail due to ADB being
unable to write to a directory. This commit addresses this additional
failure in the test by requiring the test to run as a (fake) root user
within the Bazel sandbox; the commit also updates the AvdTestRule to
mount a tmpfs to the test user's home directory. Since the test runs
within a mount namespace, this mount does not affect any other processes
outside of the mount namespace the test is running in.

As an example, when this test is running, the test will run as the root
user within the Bazel sandbox. The home directory will be /root, and the
test will mount a tmpfs to /root. This allows ADB to write files to
/root. At the end of the test, the tmpfs is unmounted from the /root
directory. No change will be made to the actual /root directory.

The test was added back to the sandbox by removing the "exclusive" tag
from the test target. Although this tag ran the test without
sandboxing, another goal of the exclusive tag was to ensure that the
test ran by itself, since the CI machine running the machine may not
have enough resources for the test to run on time. This problem was
resolved by adding the tag "cpu:100" to the test target, such that Bazel
will try to grant more resources to the test by not running other
targets simultaneously.

Bug: 120862107
Test: DebugOnEmulatorTest
Change-Id: Iabd780fcb4cc3e34b4da9808110a621f29db2151