Enable agent for AddAppCompatLibTest UI test
Enable the aspects agent to this UI test when running with bazel as a first
step to enable the agent for all UI tests. In this initial pass, we included
a very simple set of rules that simply warns when HttpRequest#openConnection
is called.

When running AddAppCompatLibTest from bazel, the agent should be attached
automatically and it will be using the default set of rules. The idea is to
do the same for all the UI tests once we verify this approach is robust and
works properly for this test in PSQ environment.

Custom agent locations can be used by specifying the aspects.agent.jar JVM
flag (when running from bazel or IDEA), and the rules file can also be
changed by providing the aspects.agent.rules flag.

Bug: 124110538
Test: AddAppCompatLibTest
Change-Id: Iabd874306e2db52a4112b3232d8c3eade69e05e7