Determine file icon lazily in "Open File or Project" dialog.
In the dialog, when you try to explore a directory with many child directories
it becomes really slow. This is because the IDE scans a list of files in a
child directory (so if you open a directory with 10 nested directory which
have 10 files each, the dialog scans 10*10 files). This scan happens only to
decide which icon to use for a directory. If we choose the icon based on the
file type, the dialog becomes much responsive (10x faster in my environment).
This problem is especially visible when you work on network file system.

This CL changes the icon selection to lazily update to the most appropriate
ones and starting from file type based icon. When you open a small directory,
you don't notice any changes before and after this CL. Ideally, this change
should be made in IntelliJ's open API code and here's the related ticket
https://youtrack.jetbrains.com/issue/IDEA-101218. Let's start with Android
plugin and work on pushing it to the core API.

Change-Id: Icb73d47ecd602bdb0f808463baaf332f24a2e529
Fixes: 37099520
Fixes: 124297091
Test: AndroidOpenFileActionTest and manual testing