Set default style settings on first launch only
There are some settings in the default profile that are
intended to be set on the first launch of Android Studio.
For example, the default right margin in the default
profile should be set to 100. At some point the code
was moved and ended up getting executed on every launch.
As a result any changes that the user made to these properties
was lost upon the next launch of Android Studio.

This fix sets a configuration property to ensure that changes
to these properties are not overwritten during subsequent launches.

Fixes: 68178350
Test: AndroidStudioInitializerTest.testModifyDefaultCodeStyleSettings
Change-Id: I9b8cb358a760b33c26d61949b44f982b31b5017f