Update test error message in GenCodeMatchTest
There was an assertion that counted a number of generated classes,
but we replace that with an explicit check of exactly which
classes we expect to get generated. This is not only more
likely to fail fast if things break in the future, but it is also
a little easier to read the intention of what the test is doing
when you know what the expected values should be.

Bug: N/A
Test: Updated
Change-Id: I9e0fac58cfbefa0545110e16c8f120d3a57fe7e8