Delete incompletely unzipped leftovers in installation temp dir.
We already delete certain things on cleanup (which is invoked
for example when operation gets cancelled), so this CL extends
that list with the directory the package is unzipped to
in the installation temp dir. That directory most likely
contains a package in an inconsistent state anyway, and
therefore would just occupy extra space for no reason.

Bug: 37092608 et al.
Test: FullInstaller currently doesn't have a test class, producing
one is TBD.
Change-Id: I1ed197eeaf5f68c16067dd113515e80ee643bd44