[Gradle Build Model] Changes to SigningConfigs
This change ensures that if the parser creates the signing config block
that it positions it at the top of the Android block so that it can be
references by other block types. This minimises the chance of the PSD
getting the build file into an incorrect state.

In this change we also fix signing configs passwords fields to ensure
they correct default to PLAIN_TEXT rather than ENVIROMENTAL_VARIABLE.

Change-Id: Ide7e7412204bdb1c0105118629491d4900b48d0b
Fixes: 122691442
Fixes: 116565314
Test: Unit tests added