Enable both legacy and new data pipeline in transport datastore
With the resource inspector and the profiler sharing the pipeline, the transport
layer should generate new pipeline events by default. The legacy DeviceProcessPoller
is now guarded by the PROFILER_UNIFIED_PIPELINE flag, so that the profilers can
still receive legacy data even if there is another party using the new pipeline.

A number of associated cleanups:
- Moved TransportServiceProxy to android-transport, since its logic is common to the pipeline.
- Moved the IDevice -> Common.Device generation logic out of TransportServiceProxy
into TransportDeviceManager. This change allows both the legacy and new event pipeline
to share the same id whether the device is represented as a stream or device. It also
allows for some cleanup as the datastore no longer have to wait for the proxy layer
to return the Common.Device to map it to a particular channel.
- Added a test constant STREAM object in DataStorePollerTest that is shared by
its subclasses. A number of test xml files are updated due to this.

Bug: 123941356
Test: Updated

Change-Id: Id4125342c49dac9b7ce22069aa71bf0e138e3278