Avoid calling Project#save on the UI thread
This CL moves the call to Project#save after sync out of the UI thread.
This shows in the freezes dashboard frequently for Windows but there is
no obvious reason to call save on the UI thread.
This keeps a similar flow to before but splits the Runnable in 3. The
first and last parts are executed on the UI thread, the middle one on
the background pool.

Change-Id: I57ea2781b91bbb62fea6d412c5a25710a65ba329
Fixes: 124039295
Test: Existing tests pass.