Simplify disk space analysis for SDK operations.
Since SDK operations are now going to use the temp
directory rooted at SDK directory, it is no longer
necessary to estimate the size of the system temp
partition and is enough to analyze the SDK partition
only. This CL therefore removes unnecessary logic
and cleans up the code a bit.

Test: Added SdkUpdaterConfigurableTest.testDiskSpaceMessages
Bug: Part of 37092608 et al. - see Id02e09d7887.
Change-Id: Ibee90983ab58e24c4c11f016ffd5e830a3735408