Identify Groovy and kts build files by extension
We log if a project uses Groovy or Kotlin build files but we are doing
it by looking if the name matches build.gradle or build.gradle.kts
completely. Gradle allows to use custom build files and if a project
uses them then we will not identify the type correctly.

This change uses the file extension only instead of the full file name.

Bug: 123667194
Test: N/A
Change-Id: Ia22664fccd757b234c087a38a63a0bd8fedf4b83