Refactor MergedManifest to avoid blocking calls
This is a first step towards making clear when obtaining a
MergedManifest might block. It also renames MergedManifest to
MergedManifestSnapshot to make obvious that you obtain the merged
version at a point on time.
The new API also allows to obtain a cached if it exists without
triggering a full reparse.

Bug: 123339491
Test: Existing tests pass
Change-Id: I4d6ffe4c978ae67c369a4f55e711c951838ea6c9