Don't navigate to embedded framework resources from Android annotations
XML editor does not recognize "http://schemas.android.com/apk/res/android" and is not capable of showing preview when the opened file is not recognized as part of an SDK, library, or project. For framework resources it means that only framework resources of the target SDK can be properly previewed in XML editor.

To avoid opening an embedded framework resource intended for rendering only, which cannot be properly previewed in XML editor, we remap it to a similar framework resource belonging to the project target. This makes navigation by clicking on the image preview consistent with going to declaration on the resource reference in an attribute value. If remapping of the resource file fails, navigation by clicking on the image preview is disabled.

Test: AndroidColorAnnotatorTest
Bug: 123860195
Change-Id: I5faeea150891fd0c67cfac6fb63e5bbc81a4f709