[UI Tests] NewInstantAppModuleTest is flaky - Fix Android API combo-box
The Android API selection comboBox values are loaded
twice. First its pre-loaded with a list of Default API levels,
and then a second list is asynchronously loaded from
the Android SDK, containing no default and third party API levels.

If loading from SDK is slow, it is possible that the user selects
an API level in the comboBox before the reply is received and
the current selection is the over-written.
This CL tries to maintain the select item, even if the list of API's
is updated.

Bug: 123684808
Test: NewInstantAppModuleTest
Change-Id: Ic5a1dcaa1826947386fd3926a30487dcd8e42af8