Improve error messages in AvdTestRule exception
AvdTestRule sometimes fails unexpectedly in our CI system. We expect the
test to run with fake root privileges inside its user namespace, but the
test does not. The username of the user should be printed in the
Exception's error message in case the test does not end up running as
root again.

Bug: 120862107
Test: N/A
Change-Id: I7b9d0f0ff68648ec8fc06bd681c5983f71c50998