Dissociate ExecutionTarget from RunConfigurations.
ExecutionTargets were unnecessarily being associated with specific
RunConfigurations when all that's needed were the application IDs.

Now whomever needs to check if an application is running on a device,
they will need to pass in the approriate application ID. This also
makes ExecutionTargets unaware of specific applications at creation time.

Bug: N/A
Test: N/A
Change-Id: I1283cba0c1fbe40f9f924e38bb93f95e3a4d72e3