Remove StreamId wrapper from transport datastore
This is a follow-up based on a comment from 2b4e6604c0fd5ab9f1babbba1d0e85cb99c6eba2.
The wrapper proves to be confusing and did not get wide adoption, which actually made
the codebase less readable and consistent. The issue became more apparent when the
pipeline was moved from using device id to stream id.

Bug: 123941356
Test: Updated
Change-Id: I3e66e0ea08934e2b37ffa25c789d7ae33c4a48d8