[NPW] Allow creating modules in sub folder - Supply values to templates
Allow module names of format :libs:lib1

Templates are being injected (TemplateValueInjector) with wrong values for
"Project Path" and "Module Name" (ATTR_PROJECT_LOCATION and ATTR_MODULE_NAME)

Instead of assuming that "Project Path" is the "Module Directory Parent" and
"Module Name" is "Module Directory Name", this CL re-factors
TemplateValueInjector@setModuleRoots to be specified directly with a
"Project Location" and "Module Name".

This CL does not fix the problem yet, it just re-factors the code to provide
the right data to the Wizard Templates.

Bug: 124359884
Test: TemplateTest and UI Tests
Change-Id: I0080fe299a2cb0a92b6c1fd0ec7e0960edfff550