Copy test project for custom Lint checks
This copies part of the project used in the Gradle integration test for
custom Lint checks, so that we can have an integration test on the IDE
side as well (to make sure that custom Lint checks show up in the IDE).

This is a direct copy (minus some removed files);
modifications will come in a subsequent commit.

Test: n/a
Bug: n/a
Change-Id: Id8412aeefbaf666275ee984f6f74d37b4c82256f