Refactor native symbolize resolution from perfd-Host to profilers
This change moves the native symbolizer logic out of datastore into IdeProfilerServices which
allows us to sever ProfilerService's dependency on a project in the future. To avoid introducing
an additional dependency from the profilers module to native-symbolizer, a wrapper interface
called NativeFrameSymbolizer is introduced for profilers-specific usage.

To avoid making redundant calls to symbolize when loading the JNI ref objects, a simple caching
mechnaism is introduced that is scoped to each LiveAllocationCaptureObject. We can consider scoping
the cache to the profilers session if loading performance becomes an issue. This also updates our
use cases of the NativeSymbolizer#symbolize method to use the selected session's abi instead of
the selected process, because users could be browsing old sessions during which the selected process
is invalid.

Bug: 79772836
Bug: 124526769
Test: Added SessionsManagerTest and IntelliJNativeFrameSymbolizerTest

Change-Id: I3061bf46258cbeef99ada53ffe9feb25c39e84ef