Simplify and fix setting of color of clipart launcher icons
Remove the myForegroundColor and myUseForegroundColor fields from LauncherIconGenerator and from LauncherLegacyIconGenerator and use the foreground image asset as the source of truth for the foreground color and its applicability. The complex propagation path of the foreground color that existed before this change sometimes resulted in setting the color to a wrong value.

Fix initial scaling of clipart images.

Test: LauncherIconGeneratorTest, LauncherLegacyGeneratorTest
Bug: 124356534
Change-Id: I47aaa43b0693f5846b62846725e6ec92e9d52313