Fix BR inconsistencies
Previously, we did not have coverage for sanity-testing the
contents of generated BR code, comparing it to what gets
compiled by the data binding compiler. In this CL, we have
updated GeneratedCodeMatchTest to add this.

Doing so uncovered that BR constants were being generated
incorrectly, using a Java-field-transformed version of the
original XML <variable> name.

That is, what should have been

BR {
   ...
   snake_case_var,
   ...
}

was being generated in PSI as

BR {
   ...
   snakeCasVar,
   ...
}

Additionally, in order to actually find the BR class, code in the
BrClassFinder needed to be updated, since previously it was
relying on finding a target R class virtual file, which has since
been obsoleted as R classes have moved to being generated in-memory.

Bug: 123021237
Test: Updated
Change-Id: I901489ade059b9998e5117ef95ca73a45ef75b5e