Reset clip bounds for each layer
The same graphic context is reused for all the design surface layers.
If a layer forgets to reset the clip area it might cause rendering issues in other layer.

This happened with the ScreenViewLayer that wasn't drawn during one repaint cycle because
the clipping was not correct and it is relying on it to create the render image.

Test: Manually tested
Bug: 125731914
Change-Id: I6acee1d329954b260ee23fa775c527a52756f184