NlModel#addComponents shouldn't make network operations
When adding components to the NlModel, we had a flow that would analyze
gradle dependencies by establishing network connections through
IdeGoogleMavenRepository. As such operations are blocking, they shouldn't
run on the UI thread, so we should wrap them in a ProgressIndicator using
a Task.Backgroundable that will run in the pooled executor.

This CL reapplies change I9a487be22ce53e99eb933d62972725991ac84998, which
fixes an important freeze in Studio, and fixes two flaky tests that started
failing quite often after the fix was merged.

Bug: 121239170
Test: NlEditorTest#testCopyAndPaste,
      NlModelTest#testAddComponentsWithDependencyCheck,
      NlModelTest#testAddComponentsNoDependencyCheckOnMove, and the
      remaining tests in intellij.android.designer should pass.

Change-Id: I31b7c1decada4d2af6895c7e7ed0ac14f9c7f853