Start reading properties from agent.
All strings are loaded from a separate string table.
Source locations and "id" values are read as resources.
The layout source location is included for the view.
A user defined property value is a value with a source
location equal to the layout of the view.

Include a "Declared Attributes" section in the panel for all user
defined property values.

Bug: 123643052
Test: N/A
Change-Id: I37a38d06b968c9e7e3bd23f7d0a7f21ffb5f303d