Removes conflicting shortcut from UpdateRunningApplication
UpdateRunningApplication is assigned a default hotkey that conflicts with
HotswapAction on Windows. This CL manually goes through the keymap manager
and looks for existing hotkeys for HotswapAction, and removes all hotkeys
belonging to URA and conflicts with HotswapAction. It also fixes an issue
where the default HotswapAction hotkeys will always get added to Studio,
even if the user removed them.

TODO: Add UI test to ensure hotkeys keep working.

Fixes: 123544634
Test: Manual, opened Studio, ensured URA keys are unassigned when conflicting.
Change-Id: I77cb5a383200e60a83ff865ec06ace400315edf4