Add missing artifact for Windows build
Problem: The windows build "artifacts" target does not depend on
"android-installer" which the Java deployer needs at runtime.
In the current form, the deployer fails when attempting to work
with a "virgin" device since it has no "android-installer" binary
to deploy.

Solution: Add "android-installer" dependency to Windows but not
remove windows since a bug prevent NDK compilation to work there.

Bug: 126608934
Test: None
Change-Id: Ic18db1b7353069c3ad6e7d100261f209660503e7