Fixing usability issue while searching for AVD Manager.
Currently, the only keywords that are supported when searching for the
AVD Manager action are "AVD" and "manager", making AVD Manager difficult to
find via the Action Search system.

This CL adds a description to the Action, allowing the Quick Action Provider
to find the AVD Manager Action to be found via words related to virtual devices,(such as "emulator", "virtual", "device", etc...).

Fixes: 37103588
Test: opened up action finder and search for related words
Change-Id: Ia96324191add419ade2585bd6049af7dbfb11151