Expand timeouts while waiting for ADB to start
The sanity tests that use the emulator - the emulator-dependent tests -
run on a very slow machine. This can cause ADB to fail to start on time,
which causes a null reference to be returned from
AndroidDebugBridge.createBridge.

The fix here would be to expand the timeout to wait for ADB to start.
The default timeout is 5 seconds, specified in
com.androdi.ddmlib.DdmPreferences.DEFAULT_TIMEOUT. AvdTestRule changes
this timeout to 30 seconds.

Bug: 126618133
Test: DebugOnEmulatorTest
Change-Id: I297d3b2b90604b79021dc0005070d010f8403eee