Simplify TransportService multi-project support to not depend on adb versions
It turns out Studio only supports one adb version across projects. When the sdk
path is updated in one project, it forces an update to other projects as well.

Given that is the case, the TransportService is greatly simplified to work without
dependency on adb. The TransportPipelineDialog has been verified to work across
multiple projects even after switching to different sdks after this change.

Bug: 79772836
Bug: 124526769
Test: N/A
Change-Id: I33a743c0295c1b8e272f83402f528694ced38a5a