Remove typing steps from AddRemoveCppDependencyTest
The test was failing in a spot where the test tried to remove the
externalNativeBuild block from the build.gradle file through mouse and
keyboard events. This was unreliable, since the mouse did not
include the closing curly brace of the externalNativeBuild block.

This test is not concerned with the ability to remove build code from
the build.gradle file. The test is concerned with whether or not the IDE
is capable of adding the externalNativeBuild block to the Gradle file
when a native option is selected through the UI. Since this is the focus
of the test, it is acceptable for the test to remove the
externalNativeBuild block from the Gradle file programmatically to
decrease the flakiness of the test.

Bug: 123659504
Test: bazel test
//tools/adt/idea/android-uitests:AddRemoveCppDependencyTest

Change-Id: I32c3e5233d5b2008f5c53b1bfb1388d31729c9f6