Remove ADB intiialization code from TransportDeviceManager
Given that the debug bridge will always be initialized now by the
device selection UI in the toolbar, the profilers should no longer
need to start the Adb. Previously, there was also some logic to detect
when the sdk path changes, but we don't actually have to handle that case
as long as we are listening to the IDebugBridgeChangeListener and IDeviceChangeListener
events.

Also removed the IdeSdkChangeListener interface that was specifically
introduced for the profilers tools window.

Bug: 79772836
Test: Removed IdeSdksTest
Change-Id: Ia0098384c260d7e15441e86242cdccaf746a25c1