Fix LayoutInspectorTest with new deployment flow and workaround
LayoutInspectorTest was failing because it was using the old deployment
flow. This change updates the test to use the new fixture for using the
new deployment flow.

Unfortunately, the test was still failing even after the update to the
deployment flow. The layout inspector menu option was apparently removed
from the menu when the test was running within the sandbox. The menu
option was still available when running outside the sandbox. This
mystery was solved by discovering that the menu was being opened before
the IDE entered smart mode while running in the sandbox. The layout
inspector option was only available when the IDE was running in smart
mode. Apparently, the test was fast enough to open the menu before the
IDE entered smart mode, which meant the test did not see the layout
inspector option.

Normally, we expect the test to close the menu and open the menu again
to check if the option appears after a short time. However, a separate
bug was affecting the test by having the test wait for menu items to
appear rather than closing and re-opening the menu. This was reported as
http://b/126957955, which notes that change
I68af73cfaa268b1b9174689bd7ba70b526ca5b5b modified the behavior of
MenuFixture.findActionMenuItem to poll the UI for a menu item rather
than exiting early by throwing an exception. This prevented the calling
method MenuFixture.isMenuPathEnabled from closing the menu. Instead,
when the findActionMenuItem timed out, a WaitTimedOutError was thrown to
isMenuPathEnabled instead of a ComponentLookupException. This error was
passed up through the stack to fail the test.

Basically, findActionMenuItem was designed to be a quick method. It
should not subsume the responsibility of polling the UI, which is
performed in higher-level methods like
IdeFrameFixture.waitAndInvokeMenuPath.

This change addresses the change in functionality of findActionMenuItem
by adding a workaround to the test to wait for the IDE to enter smart
mode before checking the menu. The workaround can be removed once the
old behavior of MenuFixture.findActionMenuItem has been restored.

Bug: 126957955
Bug: 122375969
Test: bazel test //tools/adt/idea/android-uitests:LayoutInspectorTest
Change-Id: Ia01ee76de514e67a862cc7603c05e62afd3fb74f