Remove ModuleDisposalDataService which duplicates IDEA's own behavior.
With this change and old sync in use, when AS opens a project which
includes a non-Gradle module (presumably added to the project by opening
it in IDEA) AS will no longer automatically remove it from the project
but only report that compilation is not supported for such modules.
Any Gradle modules removed from the Gradle project are automatically
removed from the project by IDEA's infrastructure.

Test: module removal is tested by GradleSyncProjectComparisonTest
Bug: 123860513
Change-Id: I93674ff5bdbc84f79640b942227e80ab424de2ed