Dragging components to menu should not block the UI thread
When adding components to a menu through GroupDragHandler, we were
calling NlDependencyManager#addDependencies from EDT. As this method is
blocking, it might block the UI and we should call it from a separate
thread.

This CL achieves that by wrapping the call to addDependencies with a
ProgressIndicator. As a side effect, the DragHangler API was modified to
allow specifying a callback to be executed after committing a drag.

Bug: 127249013
Test: MenuTest#dragCastButtonIntoActionBar,
      MenuTest#dragSearchItemIntoActionBar, and
      MenuTest#dragMenuItemIntoActionBar

Change-Id: If1739f2fbef715ce781a1dfc82406fb4671fc9a8