Fix a leak in DynamicResourceValueRepository
The MessageBus connection lifecycle was bound to the module instead of
the repository, so none of the disposed repositories could be reclaimed
after being disposed, e.g. after refreshing the layout preview.

I also noticed MultiResourceRepository was causing the dispose() method
of child repositories to be called multiple times, so added a check for
isDisposed().

Test: Manually checked that only one instance of
  DynamicResourceValueRepository is strongly reachable after refreshing
  the preview a few times. Others may be reachable via
  Disposer.ourTree.myDisposedObjects, but I assume it's WAI.
Fixes: 126959085
Change-Id: If1a1b40275785a75e99272cfe49c15189102202c