Avoid synchronized manifest creation
The manifest merger is called under read actions and we were
synchronizing at the class level.
Someone with a read action tries to get the class monitor. At the same
time, the UI thread tries to start running a write operation, blocking
new reads.
Some other thread with the class monitor, gets to the point where the
thread needs a read lock but can not progress.

Fixes: 127308479
Test: Adding tests for manifest reading
Change-Id: I8e3aa98e0f3f0bed443000911db3d8f32de38289