Improve stability of AddKotlinTest
The test was failing for several reasons:

1. The "OK" button could take a long time to be clickable. The default
wait is 10 seconds. Judging by screenshots, it appears the button can
take longer than 10 seconds to become clickable. In this case, we
improve stability of the test by greatly expanding the timeout. This is
okay because the sanity tests are only concerned with UI flow rather
than the speed of the IDE, which is also dependent on network
conditions.

2. The UI test robot can fail to enter text into text fields. This is
another case of b/119283528. This error is fixed by having the robot
wait for the clipboard to be ready before continuing the paste operation

3. The project was syncing after the Gradle file was updated. This is
resolved by waiting for the IDE to have no background tasks running
after updating the Gradle file.

Bug: 125934368
Test: AddKotlinTest
Change-Id: I171c7583de6a626f9f9380b409efd19d43e23b8b