[Gradle Sync Merge] Implement SVS on IDEA infra
This change uses the new APIs that were added to the
GradleProjectResolverExtension to implement Single Variant Sync on top
of IDEAs sync infrastructure. We do this be providing a
AndroidExtraModelProvider which contains the logic for requesting models
from Gradle which is run as part of IDEAs ProjectImportAction.

This change also adapts the following classes from our new sync to work
with slightly different inputs so they are usable in IDEAs sync. These
are IdeaAndroidModule and IdeaSelectedVariantChooser.

This change does not use the provider so no behaviour should be affected
by the change.

Test: existing
Bug: 124662919
Change-Id: Idf6ddb3928c66322c7a921676f387d6bbf1835c6